Got rid of handlebars
