+++ a4/package.json
@@ -16,6 +16,7 @@
     "mongodb": "^6.20.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-router-dom": "^7.9.3",
     "vite-express": "*"
   },
   "devDependencies": {
