+++ a4/src/client/styles.css
@@ -1,6 +1,11 @@
 body {
     background-color: aquamarine;
-    font-family: "Dosis", sans-serif;
+  font-family: "Dosis", sans-serif;
+  margin: 0;
+  height: 100vh;              
+  display: flex;
+  justify-content: center;     
+  align-items: center;         
 }
 
 form {
@@ -50,5 +55,22 @@ table{
   margin-left: auto;
   margin-right: auto;
   display: flex;
-  align-items: flex-start;    
+  align-items: flex-start;  
 }
+
+td{
+    color: black
+}
+
+label{
+    color: black
+}
+h2 {
+    color: black
+}
+
+p {
+    color:black;
+    text-align: center;
+}
+
