--- a4/src/views/login.handlebars
@@ -1,34 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-    <title>CS4241 Assignment 3</title>
-    <meta name = "viewport" content="width=device-width, initial-scale=1">
-    <meta name="description" content="Login page for CS 4241 A3">
-    <meta charset="utf-8">
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Dosis:wght@200..800&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/purecss@3.0.0/build/pure-min.css" integrity="sha384-X38yfunGUhNzHpBaEBsWLO+A0HDYOQi8ufWDkZ0k9e0eXz/tH3II7uKZ9msv++Ls" crossorigin="anonymous">
-    <link rel="stylesheet" href ="a4/src/client/styles.css">
-</head>
-
-<body>
-    <h1>
-        Welcome to the Favorite Database!
-    </h1>
-    
-    <h2>
-        Please Login In
-    </h2>
-    <form action='/login' method='POST' class = "pure-form">
-        <label for="username">Username</label> <br>
-        <input type = "text" name = "username" id = "username"> <br><br>
-        <label for="password">Password</label> <br>
-        <input type = "password" name = "password" id = "password"> <br><br>
-        <input type = "submit" value = "submit" class = "pure-button"> <br>
-    </form>
-<p id = "message">
-    {{{msg}}}
-</p>
-</body>
-</html>
